تغییر در سرویس هاثبت نمرات

+++ service/src/main/java/com/nicico/training/dto/ClassStudentDTO.java
@@ -27,7 +27,7 @@ public class ClassStudentDTO implements Serializable {
     private String reasonsfailure;
 
     @ApiModelProperty(required = true)
-    private double score;
+    private Float score;
 
     @ApiModelProperty(required = true)
     private Long studentId;
@@ -107,6 +107,22 @@ public class ClassStudentDTO implements Serializable {
         private Integer endRow;
         private Integer totalRows;
     }
+
+//    @Getter
+//    @Setter
+//    @Accessors(chain = true)
+//    @ApiModel("ClassStudentgetStudentRq")
+//    public static class getStudent {
+//    private long id;
+//    private String personnelNo;
+//    private String nationalCode;
+//    private String personnelNo2;
+//    private String firstName;
+//    private String lastName;
+//    private Float score;
+//    private String reasonsfailure;
+//    private String scoresState;
+//    }
 
 
 }
